Added more types and something
